requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pillow
opencv-python
numpy
pandas
scikit-learn
sentence-transformers
faiss-cpu
python-dotenv
openai
xgboost
torch
torchvision
pytesseract
pdf2image
joblib
imbalanced-learn
langgraph
langgraph-checkpoint-sqlite
langchain
langchain-community
typing-extensions
groq
PyPDF2
tf-keras
